requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn

requests
openai  
python-dotenv
groq

